Move run domain helpers to api/Runs
