+++ api/Mappers/RunResponseMapper.cs
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 // SPDX-FileCopyrightText: 2026 LFM contributors
 
-using Lfm.Api.Helpers;
 using Lfm.Api.Repositories;
+using Lfm.Api.Runs;
 using Lfm.Contracts.Runs;
 
 namespace Lfm.Api.Mappers;
