+++ tests/Lfm.Api.Tests/Helpers/RunModeResolverTests.cs
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 // SPDX-FileCopyrightText: 2026 LFM contributors
 
-using Lfm.Api.Helpers;
 using Lfm.Api.Repositories;
+using Lfm.Api.Runs;
 using Xunit;
 
 namespace Lfm.Api.Tests.Helpers;
