+++ tests/Lfm.Api.Tests/RunEditabilityTests.cs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 // SPDX-FileCopyrightText: 2026 LFM contributors
 
-using Lfm.Api.Helpers;
+using Lfm.Api.Runs;
 using Xunit;
 
 namespace Lfm.Api.Tests;
